+++ R/download.R
@@ -382,9 +382,9 @@ nc_data_save <- function(data, df_db, table = "naturecounts") {
 #' nc_count(username = "sample")
 #' }
 #'
-#' # Count records with American Robins in Ontario
-#' search_species("robin")
-#' nc_count(species = 15770, region = list(statprov = "ON"), username = "sample")
+#' # Count records with house finches in Ontario
+#' search_species("house finch")
+#' nc_count(species = 20350, region = list(statprov = "ON"), username = "sample")
 #'
 #' # Count all records available in the Christmas Bird Count and Breeding Bird
 #' # Survey collections (regardless of user permissions)